requirements.txt
pyTelegramBotAPI==4.14.1
pyrogram==2.0.106
TgCrypto==1.2.5
groq==0.9.0
python-dotenv==1.0.0
